created new componemt for DraggableJobBox
